Update to .NET 6.0 and latest Corvus bits
