+++ Solutions/Corvus.CosmosClient.NewtonSoft.Json.Extensions/Corvus.CosmosClient.NewtonSoft.Json.Extensions.csproj
@@ -1,0 +1,23 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <Import Project="$(EndjinProjectPropsPath)" Condition="$(EndjinProjectPropsPath) != ''" />
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Corvus.Extensions.Newtonsoft.Json" Version="2.0.6" />
+    <PackageReference Include="Endjin.RecommendedPractices.GitHub" Version="2.1.0">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Corvus.CosmosClient.Extensions\Corvus.CosmosClient.Extensions.csproj" />
+  </ItemGroup>
+
+</Project>
