--- Solutions/Corvus.Extensions.CosmosClient/Corvus.Extensions.CosmosClient.csproj
@@ -1,46 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  
-  <Import Project="$(EndjinProjectPropsPath)" Condition="$(EndjinProjectPropsPath) != ''" />
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <RootNamespace />
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
-    <PackageDescription></PackageDescription>
-    <PackageTags></PackageTags>
-    <PackageReleaseNotes></PackageReleaseNotes>
-    <NeutralLanguage>en-GB</NeutralLanguage>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Corvus.Extensions" Version="1.1.4" />
-    <PackageReference Include="Corvus.Extensions.Newtonsoft.Json" Version="2.0.6" />
-    <PackageReference Include="Corvus.Retry" Version="1.0.2" />
-    <PackageReference Include="Endjin.RecommendedPractices.GitHub" Version="2.1.0">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.23.0" />
-    <PackageReference Include="Microsoft.Bcl.HashCode" Version="1.1.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Update="Corvus\Extensions\Cosmos\Internal\ExceptionMessages.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>ExceptionMessages.resx</DependentUpon>
-    </Compile>
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Update="Corvus\Extensions\Cosmos\Internal\ExceptionMessages.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>ExceptionMessages.Designer.cs</LastGenOutput>
-      <CustomToolNamespace>Corvus.Extensions.Cosmos.Internal</CustomToolNamespace>
-    </EmbeddedResource>
-  </ItemGroup>
-</Project>
